requirements.txt
absl-py==2.3.1
altair==5.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
boto3==1.40.67
botocore==1.40.67
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
executing==2.1.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
importlib_metadata==8.5.0; python_version < "3.10"
ipykernel==6.29.5
ipython==8.27.0
ipython_pygments_lexers==0.2.2
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
narwhals==2.10.2
nest_asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.3.3
parso==0.11.3
pillow==12.0.0
platformdirs==4.2.2
polars==1.35.1
polars-runtime-32==1.35.1
prompt_toolkit==3.0.48
protobuf==6.33.0
psutil==6.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==306;
PyYAML==6.0.3
pyzmq==26.2.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
s3transfer==0.14.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.16.0
smmap==5.0.2
stack_data==0.6.3
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.226
ultralytics-thop==2.0.18
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
zipp==3.20.2
